refactor(synapse,bench): fix cross-module import references (#428)
Update stale imports in spector-synapse (5 files) and spector-bench
(16 files) after kernel package moves:

- StorageLayout: memory. -> memory.kernel.
- CognitiveRecordLayout: memory.synapse. -> memory.kernel.layout.
- SynapticHeaderConstants: memory.synapse. -> memory.kernel.layout.

Full reactor compile verified.

Refs #428
