+++ bench/spector-bench/src/test/java/com/spectrayan/spector/bench/cognitive/ZeigarnickPropertyTest.java
@@ -20,11 +20,11 @@ import java.lang.foreign.MemorySegment;
 import java.util.List;
 
 import com.spectrayan.spector.memory.model.RecallOptions;
-import com.spectrayan.spector.memory.synapse.CognitiveRecordLayout;
-import com.spectrayan.spector.memory.synapse.CognitiveRecordLayout.CognitiveHeader;
+import com.spectrayan.spector.memory.kernel.layout.CognitiveRecordLayout;
+import com.spectrayan.spector.memory.kernel.layout.CognitiveRecordLayout.CognitiveHeader;
 import com.spectrayan.spector.memory.synapse.CognitiveScorer;
 import com.spectrayan.spector.memory.synapse.IdentityCalibration;
-import com.spectrayan.spector.memory.synapse.SynapticHeaderConstants;
+import com.spectrayan.spector.memory.kernel.layout.SynapticHeaderConstants;
 
 import net.jqwik.api.Arbitraries;
 import net.jqwik.api.Arbitrary;
@@ -45,9 +45,9 @@ class ZeigarnickPropertyTest {
 
     private static final int DIMS = 8;
 
-    // ══════════════════════════════════════════════════════════════
-    // Property 26: Unresolved → decay bucket clamped to 0
-    // ══════════════════════════════════════════════════════════════
+    // â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•
+    // Property 26: Unresolved â†’ decay bucket clamped to 0
+    // â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•
 
     /**
      * Property 26: An unresolved memory (FLAG_RESOLVED not set, not pinned) should
@@ -72,13 +72,13 @@ class ZeigarnickPropertyTest {
             for (int i = 0; i < DIMS; i++) identicalVec[i] = 0.3f;
             float importance = 5.0f;
 
-            // Record 0: UNRESOLVED (bit 5 = 0) + old → should clamp to bucket 0
+            // Record 0: UNRESOLVED (bit 5 = 0) + old â†’ should clamp to bucket 0
             byte unresolvedFlags = 0; // no resolved flag, no pinned flag
             long oldTimestamp = nowMs - ageMs;
             writeRecord(segment, layout, 0, identicalVec, importance,
                     oldTimestamp, 0, unresolvedFlags, mins, scales);
 
-            // Record 1: RESOLVED + old → normal decay
+            // Record 1: RESOLVED + old â†’ normal decay
             byte resolvedFlags = SynapticHeaderConstants.FLAG_RESOLVED;
             writeRecord(segment, layout, 1, identicalVec, importance,
                     oldTimestamp, 0, resolvedFlags, mins, scales);
@@ -112,9 +112,9 @@ class ZeigarnickPropertyTest {
         }
     }
 
-    // ══════════════════════════════════════════════════════════════
+    // â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•
     // Generators
-    // ══════════════════════════════════════════════════════════════
+    // â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•
 
     @Provide
     Arbitrary<Long> oldTimestamps() {
@@ -125,9 +125,9 @@ class ZeigarnickPropertyTest {
         );
     }
 
-    // ══════════════════════════════════════════════════════════════
+    // â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•
     // Helpers
-    // ══════════════════════════════════════════════════════════════
+    // â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•
 
     private void writeRecord(MemorySegment segment, CognitiveRecordLayout layout,
                              int index, float[] vector, float importance,
