+++ synapse/spector-synapse/src/test/java/com/spectrayan/spector/synapse/memory/UserMemoryRegistryRoutingPropertyTest.java
@@ -35,7 +35,7 @@ import org.springframework.security.core.context.SecurityContextHolder;
 
 import com.spectrayan.spector.memory.SalienceProfileProvider;
 import com.spectrayan.spector.memory.SpectorMemory;
-import com.spectrayan.spector.memory.StorageLayout;
+import com.spectrayan.spector.memory.kernel.StorageLayout;
 import com.spectrayan.spector.provider.embedding.EmbeddingProvider;
 import com.spectrayan.spector.provider.generation.LlmProvider;
 import com.spectrayan.spector.spring.autoconfigure.SpectorConfigProperties;
