+++ synapse/spector-synapse/src/test/java/com/spectrayan/spector/synapse/security/UserAccountTsidPiiAvoidancePropertyTest.java
@@ -25,7 +25,7 @@ import org.springframework.jdbc.core.simple.JdbcClient;
 import org.springframework.security.crypto.password.Pbkdf2PasswordEncoder;
 
 import com.nimbusds.jwt.SignedJWT;
-import com.spectrayan.spector.memory.StorageLayout;
+import com.spectrayan.spector.memory.kernel.StorageLayout;
 import com.spectrayan.spector.memory.id.TsidGenerator;
 import com.spectrayan.spector.synapse.config.SynapseProperties;
 import com.spectrayan.spector.synapse.config.SynapseProperties.AuthProperties;
